+++ Tests/49/Test.cpp
@@ -593,7 +593,7 @@ namespace {
             using DRT = DisjointDiscreteRange<int>;
             DRT                 dr;
             static const size_t kMax_ = Debug::IsRunningUnderValgrind () ? 1000 : 10000;
-            for (int i = 0; i < kMax_; ++i) {
+            for (size_t i = 0; i < kMax_; ++i) {
                 dr.Add (i);
             }
             VerifyTestResult (dr.SubRanges ().size () == 1);
